fix : jpa 관련 매핑 타입 변경

+++ module-service-rental-reservation/src/main/kotlin/com/raillylinker/module_service_rental_reservation/jpa_beans/db1_main/entities/Db1_RaillyLinkerCompany_RentableProductReservationStateChangeHistory.kt
@@ -24,7 +24,7 @@ class Db1_RaillyLinkerCompany_RentableProductReservationStateChangeHistory(
 
     @Column(name = "state_code", nullable = false, columnDefinition = "TINYINT UNSIGNED")
     @Comment("예약 상태 코드(0 : 관리자 결제 확인, 1 : 관리자 예약 승인, 2 : 관리자 예약 거부, 3 : 사용자 예약 취소 신청, 4 : 관리자 예약 취소 승인, 5 : 예약 취소 거부, 6: 사용자 조기반납신고, 7: 관리자 조기반납 확인, 8: 결제 환불 처리)")
-    var stateCode: Int,
+    var stateCode: Byte,
 
     @Column(name = "state_change_desc", nullable = false, columnDefinition = "VARCHAR(600)")
     @Comment("상태 변경 상세")
